added test code
